Deprecate split transactions status field
It is a left over from previous serial implementations like the helix
keyboard. The field is read and written to in split communications, but
never evaluated as the current core implementation relies on the return
value of soft_serial_transaction which is treated as a bool. Therefore
status is removed and soft_serial_transaction returns a bool now.

soft_serial_get_and_clean_status is removed as it is dead code.