Move mousekey keycodes into newly freed up keycode block
The reason for this move is because ACTION LAYER reserves 0xE0-0xFF for internal usage.
Namely, for other layer functions.  This means that LT is limited to 0x00-0xDF keycodes,
and effectively blocks mousekeys from being used.

By moving the mousekeys, this allows them to be used as part of LT.