qmk new-keyboard: refactor question() responses
question() can only return empty if default is empty. If there is a
default value, it will be used if the user doesn't specify anything.

We thus need a loop around the user_name prompt because the default can
be empty, but the only reason we can end up without user_name is when it
hasn't been provided as an argument.

The real_name question will however always have a default, and there is
no need for aloop.