Swap Escape and Caps
This patch swaps Escape and Caps (swap, inswap, toggle)

This is a squashed commit with consistent sequence of keycodes.

This is popular swap option found in GNOME teaks etc.

It worked on my QMK keyboard (handwired 4x14 ortholinear compact.)

When a host PC with GNOME tweak sets to swap Escape and Caps, I had to do
this to obtain in sane key location on my QMK keyboard while keeping
that note PC usable with vim.

Properly targeting devel branch

Signed-off-by: Osamu Aoki <osamu@debian.org>