Remove Helix/rev1
Helix alpha (helix/rev1) is the alpha version of Helix. About a dozen PCBs were created and distributed to alpha testers. Less than six months later, the official version of Helix was released.
Helix alpha has not been distributed since the official version of Helix was released.

After consulting with MakotoKurauchi, the owner of the Helix project, we have come to the conclusion that there is no need to maintain helix/rev1 anymore, so we are removing it.