Add a keycode for sharing backspace and backslash/pipe.
This adds a feature similar to Grave Escape but for the Backspace which is
useful for Tsangan/HHKB layouts with split-backspace, it lets you share the
Backspace with `\` and `|`.

Many of us have a lot of muscle memory for using the pipe where the backspace is
on a Tsagan/HHKB, this is frequently used if you're in a unix shell or
programming and reaching KC_BSLS with the pinky on a 1U key can be annoying.

We introduce QK_BSPIPE that produces a pipe when shift is held or a backslash(\)
(ascii 92) if CMD/GUI/WIN is held, this follows the same behaviour of Grave
Escape.

While here rename "shifted" to "modded" in process_grave_esc.c since it reflects
better its state.