Make the rotary encoder more responsive
The resolution 4 rotary encoder on my Keychron Q2 is missing a lot of ticks, when not all 4 pulses are detected. By putting a constraint on the default state 0xF of the encoder it is possible to catch a lot of these lost ticks and the performance of the rotary encoder improves noticeable and is reliable.