Improve ENCODER_DEFAULT_POS to recognize lost ticks
When resetting the pulses due to reaching the ENCODER_DEFAULT_POS no events were emitted, when 
Modified encoder_update to be more permissive about lost pulses when returning to the ENCODER_DEFAULT_POS  to detect if a tick has happened. E.g. on resolution 4 rotary encoder all ticks with only 1, 2 or 3 detected pulses have been discarded in the original code, but when the rotary encoder returns ENCODER_DEFAULT_POS  and pulses are present, we can be sure, that a tick has happened. This modification catches all these lost ticks with only 1,2 or 3 pulses, which results in a more reliable performance of the rotary encoder.