Fix issues with premature failure of secure locking request
Again, with the pending request, if keys are already being held, they will
trigger the !pressed event on release, ending the check prematurely.  Such
as having a request key being on a layer.  To prevent this from happening,
using the pressed event works better.

(If needed, a preprocessor option can be added to this)