info.json: correct key sequence; update key labels
The `default` keymap uses an ANSI-style key assignment (no `KC_NUHS` or `KC_NUBS`), but the layout data was represented using ISO-UK labels. This commit changes the labels to be ANSI-US in labels and layout.

Additionally, the JSON object for the Up Arrow key was out of sequence. This commit moves that JSON object to be in the correct place.