Restore delta_ thresholds for scroll events
We still need these, even with the updated interface, because directly
translating the mouse movement to scroll events makes scrolling way too
sensitive.  But we need to explicitly zero the x and y values of the resultant
mouse_report struct, otherwise we get small but visible mouse movement along
with the scroll events.