Nix: Add poetry2nix override for `pillow` to fix Darwin build
The `pillow` module for `x86_64-darwin` did not build at least in the
`macos-11` environment provided by GitHub, because the `setup.py` build
script for that module was not aware of Nix and tried to perform
autodetection of various libraries; apparently it found the `jpeg2000`
library installed through Homebrew and tried to use it, which then
resulted in a build failure.

(The `pillow` module could still be built for `x86_64-linux` even with
that build script, but the resulting module did not have most optional
features enabled, because the external libraries required for those
features were not detected properly.)

The `pillow` package in Nixpkgs has a `preConfigure` script that fixes
the above problem by specifying all include and library paths explicitly
and disabling the autodetection code to prevent using any libraries
outside the Nix store.  Add a poetry2nix override which copies that
`preConfigure` script from the package in Nixpkgs, and also copies
`propagatedBuildInputs` and `buildInputs` from there, so that the
resulting package would have the appropriate dependencies.

(The `pillow` package version in the pinned Nixpkgs revision is
significantly older than the current `pillow` version, but apparently
the `preConfigure` script from that revision is good enough to build
even the current `pillow` package.)