Add Caps Word + Combos test.
Motivated by https://github.com/qmk/qmk_firmware/pull/17240, this commit
adds a unit test of using Caps Word together with Combos.

The test checks that combo chord keys and result keys work as they
should for several combos, including overlapping combos and combos that
chord tap-hold keys. To check thoroughly, each scenario is repeated to
press the combo chord keys with a few different orders and timings.

Surprisingly, all tests of this kind that I have tried pass (on master
and develop as of 2022-06-26). Maybe the underlying problem has already
been fixed, or the real problem is a scenario that hasn't been covered
with this test?