cirque_pinnacle_gestures: Integer math circular scroll
Scroll wheel clicks are a rather coarse output, make use of lib8tion to
reduce code size and speed up the process.
atan2_8() did not produce enough resolution for correct scroll movement,
replaced with a clone of that function in higher precision.

Adjusted configuration to be easier to understand, scroll validation
criteria now expressed in pixels and angles.