Fix rolled presses with Caps Word + Auto Shift.
If Caps Word and Auto Shift are both active, then currently a rolled
press like "A down, B down, A up, B up" will incorrectly produce "Ab".
This is because Auto Shift clears the weak shift mod when the A key
is released, which in the case of a roll is *after* Caps Word has set
the weak shift mod for B when it was pressed.

This commit incorporates
https://github.com/zsa/qmk_firmware/commit/42519ed513a6f07791d6a0e6012ef8ffa4d93e9f
which fixes it. An additional test `AutoShiftRolledShiftedKeys` is
added, which would previously fail but passes with this fix.