<Update> Oneshot cancel keys
Removed Sym layer key as oneshot cancel key to better allow for use of
mod keys with symbol layer. In the previous version mod keys needed to
be hit after the Sym layer key was pressed and this was not easy to do
with the super or alt modifier keys. This change allows the mod keys to
be triggered before the Sym layer is activated.