RP2040: tentative fix for XIP lockups
Due to unpredictable spikes in execution when the XIP cache is missed
and code / data has to be fetched from the external flash the virtual
timers in ChibiOS could lockup. Until a proposed patch is merged into
ChibiOS and lands in QMK we switch to the classic periodic tick
implementation which is immune to the spike in the critical execution
path.

Co-authored-by: Ilya Zhuravlev <whatever@xyz.is>