update to keymap
relying on a layer activation pause for typing ' was a source of
consistent errors when trying to type quickly.  also reaching for ;
and : when using vim or evil was also a source of errors.  making this
swap fixes both issues and the added quick pause for the ; key is not
that bad as it usually falls at the end of a phrase or line of code.