Improve positions in RGB matrix.
The rotary encoder and the key below that are in a new column.
The rotary encoder's height is inbetween rows.
The key below is kind of off-axis and thus hard to pin down to a
specific location.

The modifer keys in the bottom row are staggered compared to the other
columns.