Fix int8_t overflow in RGB heatmap effect
The RGB heatmap effect currently has an issue where pressing the
leftmost keys on the keyboard may light up the rightmost keys on the
keyboard as well.

This is because the distance calculation for the RGB heatmap effect uses
int8_t to store the row distance and column distance. However, LED
coordinates are stored as uint8_t, so it is possible to overflow the
int8_t when calculating the row distance and column distance.

Changing the row and column distance to int16_t fixes this overflow and
resolves the original issue.