Rewrite locking in split transaction handlers
The access to the split shared memory shall be as short as possible in
order to not block any pending transaction initiated by the main half of
the keyboard. Most transaction handlers are deterministic and short in
runtime and therefore can simply take the lock for the whole duration of
the handler. Other handlers execute code that might end up calling user
code or invoke interactions with external peripherals. Those handlers
need a specialized implementation that takes a lock for the shortest
possible duration. This is done in the following commit.