rp2040: use built-in integer hardware divider and optimized i64 multiplication
The rp2040 cortex m0+ cores do not implement an integer divider itself, but
instead the rp2040 provides two on-chip hardware dividers as an extra
peripheral. To use those the GCC arm-v6 eabi integer division and modulo
built-ins have to be replaced with implementations that utilize them. An
optimized i64 multiplication implementation is provided as well.