ploopyco/trackball_mini: always update encoder
The encoder derives scroll direction based on the last known wheel
position, as such it needs up-to-date wheel state at all times
to compute the right direction.

When using opt_encoder_simple.c as the implementation, the default delay
may cause the encoder to see "impossible" states (which requires
recalibration) or states that imply an opposite direction to the actual
scrolling direction. This is very evident when the wheel is
free-spinning.

This commit reorders the code to make sure the optical encoder state is
always updated, but only apply scrolling when specified conditions
are met (ie. after debounce period, middle mouse released, etc.).

Tested with opt_encoder_simple.c, where a free-spinning wheel now
generates very smooth scrolling with no apparent glitches. This has not
been tested with the default encoder, but similar benefits should be
found there.