ChibiOS: Fix USB bus disconnect handling
usb_disconnect() previously only stopped the USB driver but didn't
disconnect the device from USB bus. This is a problem with split
keyboards that fail to enumerate on bootup but are actually connected
to a USB bus. This can happen in the case that a short `SPLIT_USB_TIMEOUT`
is used but it takes longer to enumerate with the host system correctly.
In this condition the host system would still try to enumerate with the
keyboard, as it still recognizes it as a connected device, but never
succeed as we already disabled the USB driver. This in turn prevents
some systems to boot-up correctly or at least slow down the boot process
until the device is finally rejected by the host USB stack.

A drive-by fix is the case for rebooting a keyboard without a physical
reconnect and not stopping the USB driver before starting it again -
which is a violation of ChibiOS state model for the USB driver.