RP2040: only clear rx fifo for serial pio
Previously the Full-duplex implementation would have it TX FIFO cleared
although it was still sending data, as sending data happens asynchronously
and we don't wait for all data to be send out before continuing with
the execution of the main keyboard loop. Here we would run into a race
condition where we cleared all FIFOs before starting a new transaction
while the last one was still in progress. This condition can not happen with
the RX FIFO as we wait for a reception to be fully completed before
continuing the execution, therefore it is save to fully and
unconditionally clear the RX FIFO.