Insert delay between shifted chars in send_string_with_delay
When using `send_string_with_delay`, if the string contains capital
letters, it gets sent to `send_char`, which does not insert the delay
between pressing the modifier keys needed to type the shifted letter.
This results in key codes being sent unshifted in some situations, such
as ssh connections where the keyboard is too fast for the modifier taps
to register.

This commit inesrts a delay between every modifier tap when they are
detected.