correct default keymap
The `default` keymap was seemingly written for a TKL PCB with Split Backspace support and a Tsangan bottom row, but the only layout macro currently available is 2u Backspace and a Standard bottom row.

This commit removes the Split Backspace `Delete` keycode, and adds a `KC_APP` (Context Menu) keycode to the bottom row.