Introduce event type tag for keyevent_t
The current keyevent_t time resolution was limited to an 2ms interval,
this is due to the fact that "tick" events that should only drive the
internal state machine have been generated by setting the time field to
0 and/or the row, col position to 255 before. So real key events that
should generate a reaction have been generated by using the .time =
(timer_read() | 1) pattern before. This commit introduces a proper type
tag for the keyevent_t that is evaluated instead.