Adds a crkbd revision supporting the Blok MCU
The pin mapping is different between the RP2040 used in the Blok and the
identifiers used in AVR MCUs such as the ProMicro. This commit moves the
existing pin mapping into the existing 2 revisions for the crkbd/corne
to prevent redef errors and adds new files for supporting the Blok.