Decrease LED animation frequency to improve performance
The current value `RGB_MATRIX_LED_FLUSH_LIMIT 16` causes noticeable
input latency when animated RGB profiles are activated. The latency
goes away as soon as the RGB is switched off.

This issue was introduced after refactoring in
https://github.com/qmk/qmk_firmware/pull/18640

Maybe a proper fix will be refactoring the logic and adjusting
the frequency of UART communication. This is a quick fix to make
the keyboard usable with RGB enabled.