Merge Request Changes for MR 19559
Update massdrop/ctrl/keymaps/rallyrabbit/config.h to allow for stacking on top of standard massdrop/ctrl/config.h

Update massdrop/ctrl/keymaps/rallyrabbit/rules.mk to allow for stacking on top of standard massdrop/ctrl/rule.mk

Flattened the use of flower_blooming so that C code is now in rgb_matrix_user.inc and header file is in the keymap directory (flatten the structure)

Added in proper copyright text from the MR.

rgb_matrix_user.inc includes copyright for rally rabbit for a portion and HorrorTroll for the flower blooming piece.

Updated readme.md for the keymap and code that reflects the reality of the changes made as part of the MR changes.