Enable ADC1 in the `GENERIC_PROMICRO_RP2040` board
According to docs/platformdev_rp2040.md, the `GENERIC_PROMICRO_RP2040`
board should enable the peripherals that match the Sparkfun Pro Micro
RP2040 pinout in `mcuconf.h`, but leave the corresponding options in
`halconf.h` disabled (so the options that enable the peripherals don't
actually do anything in the default state, but just enabling the options
in `halconf.h` is enough to make the Pro Micro compatible peripherals
work).  Make the ADC feature behave the same way: enable the
`RP_ADC_USE_ADC1` option in the board config, so that just enabling
`HAL_USE_ADC` would be enough to use the ADC.

After this change the `handwired/onekey/rp2040` keyboard no longer needs
the ADC configuration in `mcuconf.h`, so it is removed again.