Prevent LED flicker when connecting the keyboard
The current implementation enables LEDs and then sets mode
to manual. Because of this,
[shine firmware](https://github.com/OpenAnnePro/AnnePro2-Shine)
initially starts it own LED profiles and after processing next
message switches to manual mode where LEDs are controlled by QMK.
The effect of this is that for less than a second, keyboard
lights up all white and then switches to QMK profile.

The adjustment makes sure that no Shine profile logic is used at all.