Allow a larger int for the idle timeout
Existing code used a uint32_t for the idle timer, but the comparisons
were done with timer_read and timer_elapsed so actual comparison was
on uint16_t.  This meant the idle timer could not be more than 65535
milliseconds, or just over 1 minute.

Switched the comparisons to timer_read32 and timer_elapsed32 so that
the largest possible timeout is 4,294,967,295ms, or 136 years.  While
that would be pretty ridiculous, at least this way more reasonable
longer timeouts like 15 minutes or an hour will work.