Discuss layout macro
There is no mention or visualization of how to create a layout macro in `info.json` for split keyboards.

@fauxpark helped me tremendously in Discord by creating and providing these visualizations to aid my understanding:
https://cdn.discordapp.com/attachments/440868230475677698/1073409438712610907/image.png
https://cdn.discordapp.com/attachments/440868230475677698/1073409562373259364/image.png

I strongly feel these should be in the docs.

I think it best to include a short discussion in the [split keyboard page](https://docs.qmk.fm/#/feature_split_keyboard), explaining that row pins are automatically doubled when setting `SPLIT_KEYBOARD = yes` and that QMK views the physical layouts according to those pictures.

These are the pages I used to search for answers to the question:
https://docs.qmk.fm/#/feature_split_keyboard
https://docs.qmk.fm/#/porting_your_keyboard_to_qmk?id=layout-macros
https://docs.qmk.fm/#/hardware_keyboard_guidelines?id=ltkeyboard_namehgt