Add Encoder Map back to default keymap
per honorless#5941: "Configurator generates a JSON keymap based on what is assigned in the GUI. You can't assign encoders in that GUI (... yet...) so if there is no keyboard-level callback, the encoder won't do anything.

The default keymap is what gets compiled when someone runs qmk compile -kb keyboard -km default from the command line...and also what gets copied to a new keymap folder when somebody runs qmk new-keymap. So having encoder mapping on it is a good idea. "