Added `DISCHARGE_TIME`
Addition of a `DISCHARGE_TIME` time. This is to ensure that the ghost capacitor that is the row strobed for the ADC read, is fully drained and doesn't carry charges that deviates the readings.

Tested on multiple hardware and confirmed working and improving the readings, both in terms of stability and consistency for baseline noise.