Fix rgblight layers when animations aren't enabled
PR #18338 introduced a change that deferred rgblight_set after a call to
rgblight_set_layer_state to the next invocation of rgblight_task.

However, rgblight_task is a no-op unless RGBLIGHT_USE_TIMER is set,
which only happens automatically if an RGB animation is enabled, or if
RGBLIGHT_LAYER_BLINK is enabled.

If neither of these are enabled, rgblight_set is never called
automatically after rgblight_set_layer_state, so the LED state is never
actually set.

This commit fixes this issue by ensuring that RGBLIGHT_USER_TIMER is set
if rgblight layers are enabled.