Add RGB_MATRIX_TYPING_HEATMAP_INCREASE_STEP macro
There are a lot of configuration values for configuring the cooldown of
the heatmap. However, for fast typers, the process of heating up might
be too fast. Right now, manual changes to the codebase were necessary
which in turn require to understand it first.

By defining a macro and allowing to override it, it's now easier to
adjust this value as wanted.