`remembered_mods` arg to enable fitering the mods.
This commit adds a `uint8_t* remembered_mods` arg to
get_repeat_key_eligible_user() representing tha mods that will be
remembered with the last key. It can be modified to forget certain mods
(and perhaps with conditions about on which keys, etc.). A good use case
is forgetting capitalization by dropping Shift on letter keys.

This commit also drops the JSON object for Repeat Key as @sigprof
recommended. I originally planned to add a config option for mod
filtering, but this approach enables it in the existing callback and
simplifies away the need for a separate config option.