Atmega32U SPI driver for WS2812/SK6812 RGB leds
This MR adds two WS2812/SK6812 drives that rely on the SPI
interface, sending one bit on each serial byte. The SPI
interface is set to 8MHz, so each bit in the serial byte
lasts 125ns. Three ones are sent for a "zero bit" (thus 375ns
up and the rest of the time down), and seven ones are sent
for a "one bit" (thus 875ns up and the rest of the time down).

The first driver, ws2812_spi.c, uses polling, so the only
advantage over the bit-banging default driver is the size.

The second driver, ws2812_spi_int.c, uses interrupts,
which allows to interleave the sending of the data to the
LEDs with the execution of the normal code of QMK. This
allows to control a bigger number of LEDs without timming
problems.

Both drivers are written in assembly code. The first one,
because it's so simple that it worths it. The second one,
instead, uses assembly in the ISR to ensure that it is as
fast as possible, to stole the less possible execution time
from the main code. To simplify the maintenance, it includes
as comments the equivalent C code.