quantum/action_util.c: Use uint8_t for oneshot_layer_data
Since all bits of oneshot_layer_data are used for data encoding, the
variable must not be signed. int8_t also breaks get_oneshot_layer() on
newlib, because bit shifting on signed integers is implementation
defined, so right-shift is filled up with 1s. The only reason this
might work elsewhere are implementations filling up with 0.

closes: #20418