Update keyboards/kippersun/nanetta65/keymaps/default/keymap.c
Thanks for changing, as a beginner it's too hard for me to understand some coding rules that I don't familiar with under the case that I'm not major in Computer Science. By now this firmware can work well and I think I will make a progress after a while for studying. Sincerely.

Co-authored-by: jack <0x6a73@protonmail.com>