ploopyco: add a tiny implementation of the optical encoder
This implementation is even smaller than opt_encoder_simple.c. The
runtime portions consists of 3 lines of code with no branches or
expensive calculations.

Unlike the other two implementations, this one done away with trying to
compute the trigger thresholds and instead require one to be specified
at compile time. Sample values are left in the file for interested
users, but it is recommended to compute your own, as I only own one unit
and can't say that the values are consistent between units.

The algorithm for handling state transitions is inspired by Ben Buxton's
blogpost:
http://www.buxtronix.net/2011/10/rotary-encoders-done-properly.html

This algorithm is slightly less sensitive than opt_encoder_simple.c due
to the use of one intermediary state, but that ends up working well as
the opt_encoder_simple algorithm is too sensitive to even the slight
movements caused by lifting the finger off the wheel.