Fixed some issues with example code in drag scroll
Fixed a couple of issue with the example code that I provided. 2 issues, if you left the mouse layer while still holding drag scroll it would stay held down. Fixed that issue. In addition to that, if you move the pointing device too slowly, it wouldn't scroll. Caused by the way that the scroll values were being stored.