Make last key logic more separate from repeating.
The last key tracking in Repeat Key's implementation is useful for
things far removed from Repeat Key. Referring to APIs with "repeat_key"
in the name is out of place and potentially confusing. It also occurred
to me that future QMK features might want to share this last key logic,
for instance if the Adaptive Keys PR https://github.com/qmk/qmk_firmware/pull/14034
is revived, in which case it would help to name the APIs relating to the
last key more generically and separately from Repeat Key.

This commit renames

* get_repeat_key_eligible_user() -> remember_last_key_user()
* get_repeat_key_keycode() -> get_last_keycode()
* set_repeat_key_keycode() -> set_last_keycode()
* get_repeat_key_mods() -> get_last_mods()
* set_repeat_key_mods() -> set_last_mods()

and segments last key logic to a separate process_last_key() handler.