Refactor times inverse of sqrt 2 calculation
Diagonal mouse movement is calculated with equal x and y values times 1/sqrt(2).
The current function uses bit shifting to divide and then truncates the fractional part
to an integer, causing rounding errors.

Example, when x = -7 and y = 7, the current function will return a rounded-down value
of `4` instead of `5`, when compared to a rounded floating point multiplication with
1/sqrt(2) (0.707106781):

| x | (x * 181) >> 8 | round(x * 0.707106781) |
|---|----------------|------------------------|
|-7 |     -5         |          -5            |
| 7 |      4         |           5            |

This change uses a simple sign checking ternary operator for the numerator. If negative,
subtract half of the denominator before dividing. Otherwise add half of the denominator
before dividing. The truncated integer after division will be an accurate rounding that
matches floating point operations.