usbpd_stm32g4: Configure UCPD before disabling the dead battery mode
The UCPD peripheral initially comes up in the “dead battery” mode, in
which it supplies the Rd pull-downs on the CC1 and CC2 lines without any
explicit configuration if the DBCC pins are wired appropriately.  The
dead battery mode must be disabled to use the full functionality of
UCPD; however, disabling the dead battery mode must be done only after
the UCPD peripheral has been configured in the desired mode, so that the
transition between modes happens without going through a state with no
pull-down resistance on the CC lines.  Fix the code to respect the
required initialization order.