Fix stuck note with square wave in additive DAC
If `AUDIO_DAC_SAMPLE_WAVEFORM_SQUARE` is selected by the user, on the `dac_additive` driver, on a device where `AUDIO_DAC_OFF_VALUE` is not 0 or close to 0, `stop_note` did not work, and the last note got stuck until a new one started, at which point that note got stuck instead.

One of the comments in `platforms/chibios/drivers/audio_dac_additive.c` explains why this happened:

```c
        /* zero crossing (or approach, whereas zero == DAC_OFF_VALUE, which can be configured to anything from 0 to DAC_SAMPLE_MAX)
         * ============================*=*========================== AUDIO_DAC_SAMPLE_MAX
         *                          *       *
         *                        *           *
         * ---------------------------------------------------------
         *                     *                 *                  } AUDIO_DAC_SAMPLE_MAX/100
         * --------------------------------------------------------- AUDIO_DAC_OFF_VALUE
         *                  *                       *               } AUDIO_DAC_SAMPLE_MAX/100
         * ---------------------------------------------------------
         *               *
         * *           *
         *   *       *
         * =====*=*================================================= 0x0
         */
        if (((sample_p[s] + (AUDIO_DAC_SAMPLE_MAX / 100)) > AUDIO_DAC_OFF_VALUE) && // value approaches from below
            (sample_p[s] < (AUDIO_DAC_OFF_VALUE + (AUDIO_DAC_SAMPLE_MAX / 100)))    // or above
        ) {
```

Essentially, the code in that file tries to find the next appropriate moment to stop the DAC to avoid popping, by waiting for the next "close to zero crossing" at `AUDIO_DAC_OFF_VALUE`. Unfortunately, with the square wave as it was written, there was never such a moment: the sample value was always far (> 1/100) above or below `AUDIO_DAC_OFF_VALUE`, never near it.

This commit fixes the stuck note by having half the samples in the precomputed table be exactly equal to `AUDIO_DAC_OFF_VALUR`. On devices having `AUDIO_DAC_OFF_VALUE` at half of `AUDIO_DAC_SAMPLE_MAX`, this halves the volume of the square wave; however, square waves are plenty loud on their own.